[XHAB-11] renames files, adds mock user repository, refactors code
