+++ lib/src/presentation/scenes/splash/splash_screen.dart
@@ -1,16 +1,21 @@
 import 'package:flutter/material.dart';
+import 'package:xhabits/src/data/mock/mock_user_repository.dart';
 import 'package:xhabits/src/presentation/scenes/home/home_screen.dart';
 
-import 'splashScreenBloc.dart';
-import 'splashScreenState.dart';
+import 'splash_screen_bloc.dart';
+import 'splash_screen_state.dart';
 
 class SplashScreen extends StatefulWidget {
   @override
-  _SplashScreenState createState() => _SplashScreenState();
+  _SplashScreenState createState() => _SplashScreenState(SplashScreenBloc(
+      MockUserRepository()
+  ));
 }
 
 class _SplashScreenState extends State<SplashScreen> {
-  final _splashBloc = SplashScreenBloc();
+  final SplashScreenBloc _splashBloc;
+
+  _SplashScreenState(this._splashBloc);
 
   @override
   void initState() {
