+++ lib/src/presentation/scenes/splash/splash_screen_bloc.dart
@@ -1,21 +1,21 @@
 import 'package:rxdart/rxdart.dart';
-import 'package:xhabits/src/data/userRepository.dart';
+import 'package:xhabits/src/domain/splash_screen_use_case.dart';
 
-import 'splashScreenState.dart';
+import 'splash_screen_state.dart';
 
 class SplashScreenBloc {
   BehaviorSubject<SplashScreenState> _splashStateSubject;
   Observable<SplashScreenState> get splashStateObservable => _splashStateSubject.stream;
   Future<dynamic> get closeStream => _splashStateSubject.close();
-  UserRepository _repository;
+  SplashScreenUseCase _useCase;
 
-  SplashScreenBloc() {
+  SplashScreenBloc(SplashScreenUseCase useCase) {
     _splashStateSubject = BehaviorSubject<SplashScreenState>();
-    _repository = UserRepository();
+    _useCase = useCase;
   }
 
   void loadSplash() {
-    if(_repository.isUserSignedIn()) {
+    if(_useCase.isUserSignedIn()) {
       _splashStateSubject.sink.add(SplashScreenState(false, true));
     }
     else {
